refact: Folder organization
