+++ interfaces/Character.hpp
@@ -1,5 +1,5 @@
-#ifndef CHARACTER_HPP
-#define CHARACTER_HPP
+#pragma once
+
 #include <SFML/Graphics.hpp>
 #include <memory>
 #include <list>
@@ -43,7 +43,5 @@ public:
     void resolveCollision(std::shared_ptr<Character> other);
     virtual void doAttack(sf::Vector2f &target) = 0;
     virtual void move(float deltaTime = {}) = 0;
-    virtual sf::Sprite& getSprite() = 0; 
+    virtual sf::Sprite &getSprite() = 0;
 };
-
-#endif
