+++ interfaces/Enemy.hpp
@@ -1,5 +1,4 @@
-#ifndef ENEMY_HPP
-#define ENEMY_HPP
+#pragma once
 #include <SFML/Graphics.hpp>
 #include <memory>
 #include "RangedWeapon.hpp"
@@ -7,7 +6,6 @@
 #include "Character.hpp"
 #include "Drop.hpp"
 #include <random>
-#include <iostream>
 
 /// @brief A class that represents the other characters, known as "Enemy"
 class Enemy : public Character
@@ -28,31 +26,14 @@ public:
     float centerX;
     float centerY;
 
-    Enemy(float width, float height, float speed, int maxLife, float x, float y, float cX, float cY)
-        : Character(width, height, speed, maxLife, x, y), centerX(cX), centerY(cY)
-    {
-        shape.setFillColor(sf::Color::Red);
-        shape.setRadius(width / 2);
-        shape.setPosition(this->currentPosition);
-        weapon = std::make_shared<RangedWeapon>(5, 2, 50);
-        state = EnemyState::ALIVE;
-        std::random_device rd; std::mt19937 gen(rd());
-        std::uniform_int_distribution<> dist(0, 100);
-        if (dist(gen) >= 60)
-            drop = true;
-        else
-            drop = false;
-    };
+    Enemy(float width, float height, float speed, int maxLife, float x, float y, float cX, float cY);
 
     /// @brief Check if enemy's state is DEAD
-    bool isDead() { return this->state == EnemyState::DEAD; }
+    bool isDead();
+
     /// @brief Sets enemy's state to DEAD
-    void kill()
-    {
-        this->state = EnemyState::DEAD;
-        clockDeath = sf::Clock();
-        clockDeath.restart();
-    }
+    void kill();
+
     float getTimeSinceDeath() { return clockDeath.getElapsedTime().asSeconds(); }
     bool hasDrop() { return this->drop; }
 
@@ -62,5 +43,3 @@ public:
     void doAttack(sf::Vector2f &target) override;
     virtual sf::Sprite &getSprite() override;
 };
-
-#endif
